Minor polishing - see esp. mtcp_sys.h, for gcc 4.8

+++ trunk/src/mtcp/restore_libc.h
@@ -80,7 +80,7 @@ extern "C" {
 #ifdef __i386__
 typedef unsigned short segreg_t;
 #elif __x86_64__
-typedef unsigned int segreg_t;
+typedef unsigned short segreg_t;
 #elif __arm__
 typedef unsigned int segreg_t;
 #endif
